add hiddenBreadcrumb props in PageHeader
